[fix] allow default float values for petition variables (#1867)
---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1206643811733960

+++ server/src/api/public/__types.ts
@@ -345,7 +345,7 @@ export type CreatePetitionFieldReplyInput = {
 };
 
 export type CreatePetitionVariableInput = {
-  defaultValue: Scalars["Int"]["input"];
+  defaultValue: Scalars["Float"]["input"];
   name: Scalars["String"]["input"];
 };
 
@@ -805,7 +805,7 @@ export type Mutation = {
   createOrganization: Organization;
   /** Creates a new PDF_DOCUMENT theme on the user's organization */
   createOrganizationPdfDocumentTheme: Organization;
-  /** Create parallel. */
+  /** Create parallel */
   createPetition: PetitionBase;
   /** Creates a contactless petition access */
   createPetitionAccess: PetitionAccess;
